replace from settings to clock store some clock logic
